Move undo_post_tagging to mod/editpost because of the only occurrence
